# documentacao-api/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/usuarios/{idUsuario}":{"get":{"tags":["Usuários"],"summary":"Obter usuário por id","operationId":"cadastrarUsuario_1","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioResponseDTO"}}}}},"security":[{"token":[]}]},"put":{"tags":["Usuários"],"summary":"Atualizar usuário","operationId":"atualizarUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]},"delete":{"tags":["Usuários"],"summary":"Inativar usuário","operationId":"inativarUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/sprints/{idSprint}":{"get":{"tags":["Sprints"],"summary":"Obter Sprint","operationId":"atualizarSprint_1","parameters":[{"name":"idSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintDTO"}}}}},"security":[{"token":[]}]},"put":{"tags":["Sprints"],"summary":"Atualizar Sprint","operationId":"atualizarSprint","parameters":[{"name":"idSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintDTO"}}}}},"security":[{"token":[]}]},"delete":{"tags":["Sprints"],"summary":"Remover sprint","operationId":"deleteSprint","parameters":[{"name":"idSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}":{"get":{"tags":["Projetos"],"summary":"Obter projeto por id","operationId":"getById","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoDTO"}}}}},"security":[{"token":[]}]},"put":{"tags":["Projetos"],"summary":"Atualizar projeto","operationId":"atualizarProjeto","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoDTO"}}}}},"security":[{"token":[]}]},"delete":{"tags":["Projetos"],"summary":"Inativar projeto","operationId":"inativarProjeto","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/item-backlog-sprint/{idItemBacklogSprint}":{"get":{"tags":["Itens do Backlog da Sprint"],"summary":"Obter item do backlog da sprint","operationId":"getById_1","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogSprintDTO"}}}}},"security":[{"token":[]}]},"put":{"tags":["Itens do Backlog da Sprint"],"summary":"Atualizar item do backlog da sprint","operationId":"atualizar","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogSprintDTO"}}}}},"security":[{"token":[]}]},"delete":{"tags":["Itens do Backlog da Sprint"],"summary":"Remover item do backlog da sprint","operationId":"delete","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/item-backlog-projeto/{idItemBacklogProjeto}":{"get":{"tags":["Itens do Backlog do Projeto"],"summary":"Obter item do Backlog do Projeto","operationId":"getById_2","parameters":[{"name":"idItemBacklogProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogProjetoDTO"}}}}},"security":[{"token":[]}]},"put":{"tags":["Itens do Backlog do Projeto"],"summary":"Atualizar Item Backlog Projeto","operationId":"atualizarItemBacklogProjeto","parameters":[{"name":"idItemBacklogProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogProjetoUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogProjetoDTO"}}}}},"security":[{"token":[]}]},"delete":{"tags":["Itens do Backlog do Projeto"],"summary":"Remover Item Backlog Projeto","operationId":"deleteItemBacklogProjeto","parameters":[{"name":"idItemBacklogProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/usuarios":{"post":{"tags":["Usuários"],"summary":"Cadastrar um novo usuário","operationId":"cadastrarUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioResponseDTO"}}}}},"security":[{"token":[]}]}},"/usuarios/search":{"post":{"tags":["Usuários"],"summary":"Buscar usuário","operationId":"search","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioFilterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSearchUsuarioResponseDTO"}}}}},"security":[{"token":[]}]}},"/sprints/{idSprint}/item-backlog-sprint/search":{"post":{"tags":["Sprints"],"summary":"Buscar itens backlog da sprint","operationId":"searchItensBacklogSprint","parameters":[{"name":"idSprint","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogSprintFilterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSearchItemBacklogSprintDTO"}}}}},"security":[{"token":[]}]}},"/sprints/{idSprint}/item-backlog-projeto/{idItemBacklogProjeto}/item-backlog-sprint":{"post":{"tags":["Sprints"],"summary":"Adicionar novo item backlog a sprint","operationId":"adicionarItemBacklogSprint","parameters":[{"name":"idSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idItemBacklogProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogSprintUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogSprintDTO"}}}}},"security":[{"token":[]}]}},"/projetos":{"post":{"tags":["Projetos"],"summary":"Cadastrar um novo projeto","operationId":"cadastrarProjeto","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoDTO"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}/usuarios/{idUsuario}":{"post":{"tags":["Projetos"],"summary":"Adicionar novo usuário ao projeto","operationId":"cadastrarProjetoUsuario","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]},"delete":{"tags":["Projetos"],"summary":"remover usuário do projeto","operationId":"inativarProjetoUsuario","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}/usuarios/search":{"post":{"tags":["Projetos"],"summary":"Buscar usuários do projeto","operationId":"searchProjetoUsuario","parameters":[{"name":"idProjeto","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoUsuarioFilterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSearchUsuarioResponseDTO"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}/sprints":{"post":{"tags":["Projetos"],"summary":"Cadastrar uma nova sprint ao projeto","operationId":"cadastrarSprint","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintDTO"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}/sprints/search":{"post":{"tags":["Projetos"],"summary":"Buscar sprints do projeto","operationId":"searchSprint","parameters":[{"name":"idProjeto","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SprintFilterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSearchSprintDTO"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}/item-backlog-projeto":{"post":{"tags":["Projetos"],"summary":"Cadastrar um novo Item Backlog Projeto","operationId":"cadastrarItemBacklogProjeto","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogProjetoUpsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogProjetoDTO"}}}}},"security":[{"token":[]}]}},"/projetos/{idProjeto}/item-backlog-projeto/search":{"post":{"tags":["Projetos"],"summary":"Buscar itens do Backlog do Projeto","operationId":"searchItemBacklogProjeto","parameters":[{"name":"idProjeto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemBacklogProjetoFilterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSearchItemBacklogProjetoDTO"}}}}},"security":[{"token":[]}]}},"/projetos/search":{"post":{"tags":["Projetos"],"summary":"buscar projetos do usuário autenticado","operationId":"search_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjetoFilterDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSearchProjetoDTO"}}}}},"security":[{"token":[]}]}},"/login":{"post":{"tags":["Autenticação do usuário"],"summary":"Obter token de autenticação","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}}}},"/item-backlog-sprint/{idItemBacklogSprint}/recusar-remocao":{"post":{"tags":["Itens do Backlog da Sprint"],"summary":"Recusar remoção do item ao backlog da sprint","operationId":"recusarRemocao","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/item-backlog-sprint/{idItemBacklogSprint}/recusar-inclusao":{"post":{"tags":["Itens do Backlog da Sprint"],"summary":"Recusar inclusão do item ao backlog da sprint","operationId":"recusarInclusao","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/item-backlog-sprint/{idItemBacklogSprint}/aprovar-remocao":{"post":{"tags":["Itens do Backlog da Sprint"],"summary":"Aprovar remoção item do backlog da sprint","operationId":"aprovarRemocao","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/item-backlog-sprint/{idItemBacklogSprint}/aprovar-inclusao":{"post":{"tags":["Itens do Backlog da Sprint"],"summary":"Aprovar inclusão do item ao backlog da sprint","operationId":"aprovarInclusao","parameters":[{"name":"idItemBacklogSprint","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/usuarios/senha":{"patch":{"tags":["Usuários"],"summary":"Atualizar senha do próprio usuário logado","operationId":"atualizarSenhaUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpdateSenhaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}},"/usuarios/senha-admin":{"patch":{"tags":["Usuários"],"summary":"Atualizar senha de usuário como administrador","operationId":"atualizarSenhaUsuarioByAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpdateSenhaAdminDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"token":[]}]}}},"components":{"schemas":{"UsuarioUpsertDTO":{"type":"object","properties":{"nomeCompleto":{"type":"string"},"nomeUsuario":{"type":"string"},"senha":{"type":"string"},"email":{"type":"string"},"perfil":{"type":"string","enum":["SCRUM_MASTER","PRODUCT_OWNER","DEV_TEAM","ADMINISTRADOR"]}}},"SprintUpsertDTO":{"type":"object","properties":{"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"}}},"ProjetoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"descricao":{"type":"string"},"diasSprint":{"type":"integer","format":"int32"},"duracaoHorasDaily":{"type":"number"},"duracaoHorasPlanning":{"type":"number"},"duracaoHorasReview":{"type":"number"},"duracaoHorasRetrospective":{"type":"number"},"ativo":{"type":"boolean"}}},"SprintDTO":{"type":"object","properties":{"numero":{"type":"integer","format":"int32"},"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"},"projeto":{"$ref":"#/components/schemas/ProjetoDTO"},"id":{"type":"integer","format":"int64"}}},"ProjetoUpsertDTO":{"type":"object","properties":{"nome":{"type":"string"},"descricao":{"type":"string"},"diasSprint":{"type":"integer","format":"int32"},"duracaoHorasDaily":{"type":"number"},"duracaoHorasPlanning":{"type":"number"},"duracaoHorasReview":{"type":"number"},"duracaoHorasRetrospective":{"type":"number"}}},"ItemBacklogProjetoBasicDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"codigo":{"type":"integer","format":"int64"}}},"ItemBacklogSprintDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"responsavelRealizacao":{"$ref":"#/components/schemas/UsuarioBasicDTO"},"sprint":{"$ref":"#/components/schemas/SprintBasicDTO"},"horasEstimadas":{"type":"number"},"horasRealizadas":{"type":"number"},"itemBacklogProjeto":{"$ref":"#/components/schemas/ItemBacklogProjetoBasicDTO"},"situacao":{"type":"string","enum":["A_FAZER","DESENVOLVENDO","BLOQUEADO","PRONTO"]},"descricaoBloqueio":{"type":"string"},"prioridade":{"type":"integer","format":"int64"},"pendenteAprovacao":{"type":"boolean"},"pendenteRemocao":{"type":"boolean"},"responsavelInclusao":{"$ref":"#/components/schemas/UsuarioBasicDTO"},"dataInclusao":{"type":"string","format":"date-time"},"responsavelAprovacao":{"$ref":"#/components/schemas/UsuarioBasicDTO"},"dataAprovacao":{"type":"string","format":"date-time"}}},"SprintBasicDTO":{"type":"object","properties":{"numero":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"}}},"UsuarioBasicDTO":{"type":"object","properties":{"nomeUsuario":{"type":"string"},"id":{"type":"integer","format":"int64"}}},"ItemBacklogProjetoUpsertDTO":{"type":"object","properties":{"titulo":{"type":"string"},"descricao":{"type":"string"},"prioridade":{"type":"integer","format":"int64"},"situacao":{"type":"string","enum":["EM_ANALISE","APROVADO","CONCLUIDO"]}}},"ItemBacklogProjetoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"codigo":{"type":"integer","format":"int64"},"descricao":{"type":"string"},"prioridade":{"type":"integer","format":"int64"},"situacao":{"type":"string","enum":["EM_ANALISE","APROVADO","CONCLUIDO"]},"dataCriacao":{"type":"string","format":"date-time"},"autor":{"$ref":"#/components/schemas/UsuarioBasicDTO"},"projeto":{"$ref":"#/components/schemas/ProjetoBasicDTO"}}},"ProjetoBasicDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"}}},"UsuarioResponseDTO":{"type":"object","properties":{"nomeCompleto":{"type":"string"},"nomeUsuario":{"type":"string"},"email":{"type":"string"},"ativo":{"type":"boolean"},"perfil":{"type":"string","enum":["SCRUM_MASTER","PRODUCT_OWNER","DEV_TEAM","ADMINISTRADOR"]},"id":{"type":"integer","format":"int64"}}},"UsuarioFilterDTO":{"type":"object","properties":{"tamanhoPagina":{"type":"integer","format":"int32"},"pagina":{"type":"integer","format":"int32"},"nomeCampoOrdenacao":{"type":"string"},"direcaoOrdenacao":{"type":"string","enum":["ASC","DESC"]},"nomeCompleto":{"type":"string"},"nomeUsuario":{"type":"string"},"email":{"type":"string"},"ativo":{"type":"boolean"},"perfil":{"type":"string","enum":["SCRUM_MASTER","PRODUCT_OWNER","DEV_TEAM","ADMINISTRADOR"]}}},"ResponseSearchUsuarioResponseDTO":{"type":"object","properties":{"lista":{"type":"array","items":{"$ref":"#/components/schemas/UsuarioResponseDTO"}},"total":{"type":"integer","format":"int64"}}},"ItemBacklogSprintFilterDTO":{"type":"object","properties":{"tamanhoPagina":{"type":"integer","format":"int32"},"pagina":{"type":"integer","format":"int32"},"nomeCampoOrdenacao":{"type":"string"},"direcaoOrdenacao":{"type":"string","enum":["ASC","DESC"]},"numero":{"type":"integer","format":"int32"},"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"}}},"ResponseSearchItemBacklogSprintDTO":{"type":"object","properties":{"lista":{"type":"array","items":{"$ref":"#/components/schemas/ItemBacklogSprintDTO"}},"total":{"type":"integer","format":"int64"}}},"ItemBacklogSprintUpsertDTO":{"type":"object","properties":{"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"},"horasEstimadas":{"type":"number"},"horasRealizadas":{"type":"number"},"situacao":{"type":"string","enum":["A_FAZER","DESENVOLVENDO","BLOQUEADO","PRONTO"]},"descricaoBloqueio":{"type":"string"},"prioridade":{"type":"integer","format":"int64"}}},"ProjetoUsuarioFilterDTO":{"type":"object","properties":{"tamanhoPagina":{"type":"integer","format":"int32"},"pagina":{"type":"integer","format":"int32"},"nomeCampoOrdenacao":{"type":"string"},"direcaoOrdenacao":{"type":"string","enum":["ASC","DESC"]},"nomeUsuario":{"type":"string"},"perfil":{"type":"string","enum":["SCRUM_MASTER","PRODUCT_OWNER","DEV_TEAM","ADMINISTRADOR"]},"ativo":{"type":"boolean"}}},"SprintFilterDTO":{"type":"object","properties":{"tamanhoPagina":{"type":"integer","format":"int32"},"pagina":{"type":"integer","format":"int32"},"nomeCampoOrdenacao":{"type":"string"},"direcaoOrdenacao":{"type":"string","enum":["ASC","DESC"]},"numero":{"type":"integer","format":"int32"},"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"}}},"ResponseSearchSprintDTO":{"type":"object","properties":{"lista":{"type":"array","items":{"$ref":"#/components/schemas/SprintDTO"}},"total":{"type":"integer","format":"int64"}}},"ItemBacklogProjetoFilterDTO":{"type":"object","properties":{"tamanhoPagina":{"type":"integer","format":"int32"},"pagina":{"type":"integer","format":"int32"},"nomeCampoOrdenacao":{"type":"string"},"direcaoOrdenacao":{"type":"string","enum":["ASC","DESC"]},"titulo":{"type":"string"},"codigo":{"type":"integer","format":"int64"},"situacao":{"type":"string","enum":["EM_ANALISE","APROVADO","CONCLUIDO"]},"idAutor":{"type":"integer","format":"int64"}}},"ResponseSearchItemBacklogProjetoDTO":{"type":"object","properties":{"lista":{"type":"array","items":{"$ref":"#/components/schemas/ItemBacklogProjetoDTO"}},"total":{"type":"integer","format":"int64"}}},"ProjetoFilterDTO":{"type":"object","properties":{"tamanhoPagina":{"type":"integer","format":"int32"},"pagina":{"type":"integer","format":"int32"},"nomeCampoOrdenacao":{"type":"string"},"direcaoOrdenacao":{"type":"string","enum":["ASC","DESC"]},"nome":{"type":"string"}}},"ResponseSearchProjetoDTO":{"type":"object","properties":{"lista":{"type":"array","items":{"$ref":"#/components/schemas/ProjetoDTO"}},"total":{"type":"integer","format":"int64"}}},"AuthenticationDTO":{"type":"object","properties":{"nomeUsuario":{"type":"string"},"senha":{"type":"string"}}},"LoginResponseDTO":{"type":"object","properties":{"token":{"type":"string"}}},"UsuarioUpdateSenhaDTO":{"type":"object","properties":{"senhaAntiga":{"type":"string"},"novaSenha":{"type":"string"}}},"UsuarioUpdateSenhaAdminDTO":{"type":"object","properties":{"novaSenha":{"type":"string"},"id":{"type":"integer","format":"int64"}}}},"securitySchemes":{"token":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}